fixed all bugs

+++ ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: hzakharc <hzakharc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/06 22:10:37 by hzakharc          #+#    #+#             */
-/*   Updated: 2024/06/06 22:10:38 by hzakharc         ###   ########.fr       */
+/*   Updated: 2024/06/10 13:01:49 by hzakharc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	char	*str;
-	size_t	i;
+	char			*str;
+	unsigned int	i;
 
 	i = 0;
 	str = ft_strdup(s);
@@ -29,3 +29,20 @@ char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 	str[i] = '\0';
 	return (str);
 }
+
+// #include <stdio.h>
+
+// char	test_func(unsigned int i, char c)
+// {
+// 	return (c + i);
+// }
+
+// int main(void)
+// {
+// 	char *str = "hello";
+// 	printf("Before: %s\t", str);
+// 	char *res = ft_strmapi(str, test_func);
+// 	printf("After: %s\n", res);
+
+// 	return (0);
+// }
